Rename index file to avoid eslint errors
